Update jVectorMap to fix various tooltip problems
Change jVectorMap to reuse a single div as a tooltip rather than
creating a new one everytime, which sometimes led to extra tooltips
visible that never went away. Also move the tooltip div inside the
jvectormap-container so that when it is removed (like changing Kibana
conditions to something with no data), the tooltips go away as well.